create AbstractPricingRule to integrate methods on Rule classes
