+++ src/domain/PricingRules/GeneralMemberRule.php
@@ -6,25 +6,15 @@ namespace CinemaTicketPricing\PricingRules;
 use CinemaTicketPricing\TicketPrice;
 use CinemaTicketPricing\TicketPriceDeterminants;
 
-class GeneralMemberRule implements PricingRuleInterface
+class GeneralMemberRule extends AbstractPricingRule implements PricingRuleInterface
 {
-    private $determinants;
-
-    private $prices;
-
     public function __construct(TicketPriceDeterminants $determinants, array $prices)
     {
-        $this->determinants = $determinants;
-        $this->prices       = $prices;
+        parent::__construct($determinants, $prices);
     }
 
     public function match(): bool
     {
         return true;
-    }
-
-    public function value(): TicketPrice
-    {
-        return new TicketPrice(1000);
     }
 }
