+++ src/domain/PricingRules/RegularRule.php
@@ -6,28 +6,15 @@ use CinemaTicketPricing\TicketPrice;
 use CinemaTicketPricing\TicketPriceDeterminants;
 use CinemaTicketPricing\PriceFinder;
 
-class RegularRule implements PricingRuleInterface
+class RegularRule extends AbstractPricingRule implements PricingRuleInterface
 {
-    private $determinants;
-
-    private $prices;
-
     public function __construct(TicketPriceDeterminants $determinants, array $prices)
     {
-        $this->determinants = $determinants;
-        $this->prices       = $prices;
+        parent::__construct($determinants, $prices);
     }
 
     public function match(): bool
     {
         return true;
-    }
-
-    public function value(): TicketPrice
-    {
-        $finder = new PriceFinder($this->determinants);
-        $price = $finder->find($this->prices);
-        // finder->find で数値を返す必要あり
-        return new TicketPrice($price);
     }
 }
